Handle rate limiting, separate concerns
Rate limiting: discord has channel-specific rate-limiting that needs to
be observed and handled.

Separation of concerns: there was a set of core interactions with the
API and some scripting to use that... debugging was difficult while
everything was all lumped together, so separated them out so I could
more easily test functions in the small (see `scripts/debug.js`).